refactor(cli): extract cockpit parser args
